Refactor the naming of files to project name and add author tags for all methods
