+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.spendswift.SpendSwift")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("spendswift")
     archiveClassifier.set("")
 }
