+++ seedu/spendswift/CategoryManager.class
