+++ src/main/java/seedu/spendswift/command/ExpenseManager.java
@@ -1,8 +1,8 @@
 //@@author glenda-1506
-package seedu.duke.command;
+package seedu.spendswift.command;
 
-import seedu.duke.Format;
-import seedu.duke.parser.InputParser;
+import seedu.spendswift.Format;
+import seedu.spendswift.parser.InputParser;
 
 import java.util.ArrayList;
 import java.util.HashMap;
