+++ src/main/java/seedu/spendswift/command/TrackerData.java
@@ -1,5 +1,6 @@
-package seedu.duke.command;
+package seedu.spendswift.command;
 
+//@@author kq2003
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
