+++ src/test/java/seedu/spendswift/DukeMethodTest.java
